refactor: move mongodb file to persistense folder
