+++ src/AuthApi.Application/Infrastructure/Persistence/MongoDBDatabaseMapper.cs
@@ -1,6 +1,6 @@
-﻿using AuthApi.Application.Infrastructure.Mappings;
+﻿using AuthApi.Application.Infrastructure.Persistence.Mappings;
 
-namespace AuthApi.Application.Infrastructure;
+namespace AuthApi.Application.Infrastructure.Persistence;
 
 public static class MongoDBDatabaseMapper
 {
